social_media_api/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==42.0.8
Django==5.1.1
djangorestframework==3.15.2
ggshield==1.31.0
idna==3.10
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
mypy-extensions==1.0.0
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
platformdirs==3.0.0
pycparser==2.22
pygitguardian==1.16.0
Pygments==2.18.0
PyJWT==2.6.0
python-dotenv==0.21.1
PyYAML==6.0.2
requests==2.32.3
rich==12.5.1
setuptools==75.1.0
sqlparse==0.5.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
